pronoteChatGPT/requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
WTForms==3.1.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.0